chore(interface-types) Split the `serde` module…
… into `serde/ser.rs` and `serde/de.rs`. I like smaller files.
